Feat/SA2FTA-60/Vote

Approved-by: baniulis.tomas
Approved-by: Agne Moleikaityte

+++ src/main/java/com/sourcery/oirs/database/entity/IssueEntity.java
@@ -23,5 +23,6 @@ public class IssueEntity {
     private Timestamp finishTime;
     private UUID employeeId;
     private UUID officeId;
-    private Double rating;
-}
+    private int rating;
+
+    }
